将ClassTime单独写成类
